Carry death count in FieldAddUser for tombstone collision
The reference writes the dead player's DeathCount in the FieldAddUser
character block (right after current/max health) plus the proxy dead flag;
the client uses the death count to build the targetable/collidable tombstone.
ms2ex sent the dead flag but wrote 0 for the death count, so a joining player
rendered the tombstone without a hit box.

- CharacterList.put_character/3 accepts a death_count (default 0 for login
  list / party); FieldAddUser forwards the peer's count.
- skip dead players in the periodic proxy position broadcast so a
  position-only update doesn't clear their dead/collision state on clients.
- extract death/revive logic into Character.Revival.
- ROADMAP: tombstone-hit-by-others is still broken in-field; add next-step
  note to compare against a live capture of the reference death sequence.
